refactor presence channel into own namespace
